[IMP] Separate each class in its own file
