+++ l10n_br_nfe/models/__init__.py
@@ -3,6 +3,9 @@
 from . import res_partner
 from . import res_company
 from . import document
+from . import document_line
+from . import res_city
+from . import res_country_state
 from . import res_config_settings
 from . import cfop
 from . import document_cancel
