+++ l10n_br_nfe/models/res_city.py
@@ -1,11 +1,10 @@
-# Copyright 2019 Akretion (Raphaël Valyi <raphael.valyi@akretion.com>)
+# Copyright 2020 KMEE INFORMATICA LTDA
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
-
 from odoo import api, models
 
 
 class ResCity(models.Model):
-    _inherit = 'res.city'
+    _inherit = "res.city"
     _nfe_search_keys = ['ibge_code']
 
     # TODO understand why this is still required
