+++ l10n_br_nfe/models/uom_uom.py
@@ -1,11 +1,10 @@
-# Copyright 2019 Akretion (Raphaël Valyi <raphael.valyi@akretion.com>)
+# Copyright 2020 KMEE INFORMATICA LTDA
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
-
 from odoo import api, models
 
 
 class Uom(models.Model):
-    _inherit = 'uom.uom'
+    _inherit = "uom.uom"
 
     @api.model
     def match_or_create_m2o(self, rec_dict, parent_dict,
